// frontend/apps/web/src/domains/api/generated/model/billingStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Applark
 * OpenAPI spec version: 0.1.0
 */
import type { BillingStatusAccess } from './billingStatusAccess';
import type { BillingStatusCosts } from './billingStatusCosts';

export interface BillingStatus {
  tier: string;
  credits: number;
  access: BillingStatusAccess;
  costs: BillingStatusCosts;
}